pyproject.toml
[tool.poetry]
name = "6s_software_doc_design"
version = "0.1.0"
description = ""
authors = ["smallpoxlviv <walterwhite.99.100@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.2"
Flask-RESTful = "^0.3.9"
SQLAlchemy = "^1.4.37"
mysqlclient = "^2.1.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
